+++ packages/components/src/Shared/Modal/ModalByEntityDomain.tsx
@@ -2,6 +2,7 @@ import { computeDomain } from "@utils/computeDomain";
 import type { AllDomains } from "@core";
 import type { ModalProps } from "./";
 import { ModalLightControls } from "./ModalLightControls";
+import { ModalClimateControls } from "./ModalClimateControls";
 
 export interface ModalByEntityDomainProps extends Omit<ModalProps, "children"> {
   entity: `${AllDomains}.${string}`;
@@ -17,6 +18,10 @@ export function ModalByEntityDomain({
       return (
         <ModalLightControls entity={entity as `light.${string}`} {...rest} />
       );
+    case "climate":
+        return (
+          <ModalClimateControls entity={entity as `climate.${string}`} {...rest} />
+        );
     default:
       return null;
   }