+++ packages/core/src/types/entity/index.ts
@@ -1,7 +1,15 @@
-import { LightEntityAttributes } from "./light";
+import type { LightEntityAttributes } from "./light";
+import type { HvacMode, ClimateEntityAttributes } from "./climate";
+
+export * from "./light";
+export * from "./climate";
 
 export interface DefinedPropertiesByDomain {
   ["light"]: {
     attributes: LightEntityAttributes;
+  };
+  ["climate"]: {
+    attributes: ClimateEntityAttributes;
+    state: HvacMode;
   };
 }